Adding ToStreamIterator for LINQ
